7.1 Переносит константы, утилиты, ActionCreator согласно критериям
