+++ src/const.js
@@ -1,1 +1,12 @@
 export const ALL_GENRES = `All genres`;
+
+export const MAX_SIMILAR_MOVIES_NUMBER = 4;
+
+export const MAX_GENRES_NUMBER = 10;
+
+
+export const TabName = {
+  OVERVIEW: `overview`,
+  DETAILS: `details`,
+  REVIEWS: `reviews`
+};
